Adding CloudFoundryEntity
