+++ src/main/java/org/apache/brooklyn/cloudfoundry/entity/VanillaPaasApplicationCloudFoundryDriver.java
@@ -19,8 +19,6 @@
 package org.apache.brooklyn.cloudfoundry.entity;
 
 
-import static com.google.common.base.Preconditions.checkNotNull;
-
 import java.io.File;
 import java.io.IOException;
 import java.net.HttpURLConnection;
@@ -42,31 +40,24 @@ import org.slf4j.LoggerFactory;
 
 import com.google.common.collect.ImmutableList;
 
-public class VanillaPaasApplicationCloudFoundryDriver implements VanillaPaasApplicationDriver {
+public class VanillaPaasApplicationCloudFoundryDriver extends EntityPaasCloudFoundryDriver
+        implements VanillaPaasApplicationDriver {
 
     public static final Logger log = LoggerFactory
             .getLogger(VanillaPaasApplicationCloudFoundryDriver.class);
 
-    private final CloudFoundryPaasLocation location;
-    private VanillaCloudFoundryApplicationImpl entity;
     private String applicationName;
     private String applicationUrl;
 
     public VanillaPaasApplicationCloudFoundryDriver(VanillaCloudFoundryApplicationImpl entity,
                                                     CloudFoundryPaasLocation location) {
-        this.entity = checkNotNull(entity, "entity");
-        this.location = checkNotNull(location, "location");
+        super(entity, location);
         applicationName = entity.getApplicationName();
     }
 
     @Override
     public VanillaCloudFoundryApplicationImpl getEntity() {
-        return entity;
-    }
-
-    @Override
-    public CloudFoundryPaasLocation getLocation() {
-        return location;
+        return super.getEntity();
     }
 
     @Override
@@ -78,13 +69,13 @@ public class VanillaPaasApplicationCloudFoundryDriver implements VanillaPaasAppl
     }
 
     private String deploy() {
-        Map<String, Object> params = MutableMap.copyOf(entity.config().getBag().getAllConfig());
+        Map<String, Object> params = MutableMap.copyOf(getEntity().config().getBag().getAllConfig());
         params.put(VanillaCloudFoundryApplication.APPLICATION_NAME.getName(), applicationName);
         if (!Strings.isBlank(VanillaCloudFoundryApplication.ARTIFACT_PATH.getName())) {
             params.put(VanillaCloudFoundryApplication.ARTIFACT_PATH.getName(), getLocalPath());
         }
 
-        applicationUrl = location.deploy(params);
+        applicationUrl = getLocation().deploy(params);
         return applicationUrl;
     }
 
@@ -104,7 +95,7 @@ public class VanillaPaasApplicationCloudFoundryDriver implements VanillaPaasAppl
     }
 
     private DownloadResolver getDownloadResolver() {
-        String artifactUrl = entity.getConfig(VanillaCloudFoundryApplication.ARTIFACT_PATH);
+        String artifactUrl = getEntity().getConfig(VanillaCloudFoundryApplication.ARTIFACT_PATH);
         return new BasicDownloadResolver(ImmutableList.of(artifactUrl),
                 FileNameResolver.findArchiveNameFromUrl(artifactUrl));
     }
@@ -114,55 +105,55 @@ public class VanillaPaasApplicationCloudFoundryDriver implements VanillaPaasAppl
     }
 
     protected void configureEnv() {
-        setEnv(entity.getConfig(VanillaCloudFoundryApplication.ENV));
+        setEnv(getEntity().getConfig(VanillaCloudFoundryApplication.ENV));
     }
 
     @Override
     public void setEnv(Map<String, String> env) {
         if ((env != null) && (!env.isEmpty())) {
-            location.setEnv(applicationName, env);
+            getLocation().setEnv(applicationName, env);
         }
-        entity.sensors().set(VanillaCloudFoundryApplication.ENV,
-                location.getEnv(applicationName));
+        getEntity().sensors().set(VanillaCloudFoundryApplication.ENV,
+                getLocation().getEnv(applicationName));
     }
 
     private void launch() {
-        location.startApplication(applicationName);
+        getLocation().startApplication(applicationName);
     }
 
     private void postLaunch() {
-        entity.sensors().set(Attributes.MAIN_URI, URI.create(applicationUrl));
-        entity.sensors().set(VanillaCloudFoundryApplication.ROOT_URL, applicationUrl);
-        updateMemorySensor(location.getMemory(applicationName));
-        updateDiskSensor(location.getDiskQuota(applicationName));
-        updateInstancesSensor(location.getInstancesNumber(applicationName));
+        getEntity().sensors().set(Attributes.MAIN_URI, URI.create(applicationUrl));
+        getEntity().sensors().set(VanillaCloudFoundryApplication.ROOT_URL, applicationUrl);
+        updateMemorySensor(getLocation().getMemory(applicationName));
+        updateDiskSensor(getLocation().getDiskQuota(applicationName));
+        updateInstancesSensor(getLocation().getInstancesNumber(applicationName));
     }
 
     private void updateMemorySensor(int memory) {
-        entity.sensors().set(VanillaCloudFoundryApplication.ALLOCATED_MEMORY, memory);
+        getEntity().sensors().set(VanillaCloudFoundryApplication.ALLOCATED_MEMORY, memory);
     }
 
     private void updateDiskSensor(int disk) {
-        entity.sensors().set(VanillaCloudFoundryApplication.ALLOCATED_DISK, disk);
+        getEntity().sensors().set(VanillaCloudFoundryApplication.ALLOCATED_DISK, disk);
     }
 
     private void updateInstancesSensor(int instances) {
-        entity.sensors().set(VanillaCloudFoundryApplication.INSTANCES, instances);
+        getEntity().sensors().set(VanillaCloudFoundryApplication.INSTANCES, instances);
     }
 
     @Override
     public void restart() {
-        location.restartApplication(applicationName);
+        getLocation().restartApplication(applicationName);
     }
 
     @Override
     public void stop() {
-        location.stopApplication(applicationName);
+        getLocation().stopApplication(applicationName);
     }
 
     @Override
     public void delete() {
-        location.deleteApplication(applicationName);
+        getLocation().deleteApplication(applicationName);
     }
 
     @Override
@@ -172,20 +163,20 @@ public class VanillaPaasApplicationCloudFoundryDriver implements VanillaPaasAppl
 
     @Override
     public void setMemory(int memory) {
-        location.setMemory(applicationName, memory);
-        updateMemorySensor(location.getMemory(applicationName));
+        getLocation().setMemory(applicationName, memory);
+        updateMemorySensor(getLocation().getMemory(applicationName));
     }
 
     @Override
     public void setDiskQuota(int diskQuota) {
-        location.setDiskQuota(applicationName, diskQuota);
-        updateDiskSensor(location.getDiskQuota(applicationName));
+        getLocation().setDiskQuota(applicationName, diskQuota);
+        updateDiskSensor(getLocation().getDiskQuota(applicationName));
     }
 
     @Override
     public void setInstancesNumber(int instances) {
-        location.setInstancesNumber(applicationName, instances);
-        updateInstancesSensor(location.getInstancesNumber(applicationName));
+        getLocation().setInstancesNumber(applicationName, instances);
+        updateInstancesSensor(getLocation().getInstancesNumber(applicationName));
     }
 
     public boolean isRunning() {
